chore: restructure and make it compile
