+++ indexing/src/embedders/mod.rs
@@ -1,0 +1,3 @@
+pub mod openai;
+
+pub type Embeddings = Vec<Vec<f32>>;
