+++ infrastructure/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2021"
 anyhow = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-async-openai = { workspace = true }
 async-trait = { workspace = true }
 qdrant-client = { version = "1.9.0" }
 dotenvy = "0.15.7"
